# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Abit service.",
        "title": "Abit IS",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/semester": {
            "get": {
                "description": "get a list of all semesters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "semester"
                ],
                "summary": "List all semesters",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Semester"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing semester by semester number",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "semester"
                ],
                "summary": "Update an existing semester",
                "parameters": [
                    {
                        "description": "Semester object",
                        "name": "semester",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Semester"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Semester"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new semester to the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "semester"
                ],
                "summary": "Create a new semester",
                "parameters": [
                    {
                        "description": "Add semester",
                        "name": "semester",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Semester"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/v1/semester/{num}": {
            "get": {
                "description": "Get details of a semester by semester number",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "semester"
                ],
                "summary": "Get details of a semester",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Semester Number",
                        "name": "num",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Semester"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a semester by semester number",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "semester"
                ],
                "summary": "Delete a semester",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Semester Number",
                        "name": "num",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/v1/study_plan": {
            "get": {
                "description": "get current study plan",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "study_plan"
                ],
                "summary": "Get current study plan",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.StudyPlan"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing study plan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "study_plan"
                ],
                "summary": "Update current study plan",
                "parameters": [
                    {
                        "description": "Study Plan",
                        "name": "study_plan",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.StudyPlan"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.StudyPlan"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.EduForm": {
            "type": "string",
            "enum": [
                "Extramural",
                "Full-Time"
            ],
            "x-enum-varnames": [
                "EduFormExtramural",
                "EduFormFullTime"
            ]
        },
...
        "model.Season": {
            "type": "string",
            "enum": [
                "Fall",
                "Spring"
            ],
            "x-enum-varnames": [
                "SeasonFall",
                "SeasonSpring"
            ]
        },
        "model.Semester": {
            "type": "object",
            "properties": {
                "season": {
                    "$ref": "#/definitions/model.Season"
                },
                "semester_num": {
                    "type": "integer"
                }
            }
        },
        "model.StudyPlan": {
            "type": "object",
            "properties": {
                "cost": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "education_form": {
                    "$ref": "#/definitions/model.EduForm"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "years": {
                    "type": "integer"
                }
            }
        },
...
    }
}